# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "FavoriteBatchUpdate": {
        "properties": {
          "action": {
            "default": "add",
            "enum": [
              "add",
              "remove"
            ],
            "title": "Action",
            "type": "string"
          },
          "file_ids": {
            "items": {
              "type": "integer"
            },
            "title": "File Ids",
            "type": "array"
          }
        },
        "required": [
          "file_ids"
        ],
        "title": "FavoriteBatchUpdate",
        "type": "object"
      },
      "FavoriteTagComboCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "tags"
        ],
        "title": "FavoriteTagComboCreate",
        "type": "object"
      },
      "FavoriteTagComboInfo": {
        "properties": {
          "added_at": {
            "title": "Added At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "tags",
          "added_at"
        ],
        "title": "FavoriteTagComboInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Keivotos - Danbooru",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/blacklist-tags/{tag_name}": {
      "delete": {
        "operationId": "remove_blacklist_tag_api_blacklist_tags__tag_name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tag_name",
            "required": true,
            "schema": {
              "title": "Tag Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Blacklist Tag"
      },
      "post": {
        "operationId": "add_blacklist_tag_api_blacklist_tags__tag_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "tag_name",
            "required": true,
            "schema": {
              "title": "Tag Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Blacklist Tag"
      }
    },
...
    "/api/favorite-tag-combos": {
      "get": {
        "operationId": "list_favorite_tag_combos_api_favorite_tag_combos_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FavoriteTagComboInfo"
                  },
                  "title": "Response List Favorite Tag Combos Api Favorite Tag Combos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Favorite Tag Combos"
      },
      "post": {
        "operationId": "create_favorite_tag_combo_api_favorite_tag_combos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteTagComboCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteTagComboInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Favorite Tag Combo"
      }
    },
    "/api/favorite-tag-combos/{combo_id}": {
      "delete": {
        "operationId": "delete_favorite_tag_combo_api_favorite_tag_combos__combo_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "combo_id",
            "required": true,
            "schema": {
              "title": "Combo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Favorite Tag Combo"
      }
    },
...
    "/api/favorite-tags/{tag_name}": {
      "post": {
        "operationId": "toggle_favorite_tag_api_favorite_tags__tag_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "tag_name",
            "required": true,
            "schema": {
              "title": "Tag Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Favorite Tag"
      }
    },
    "/api/favorite-tags/{tag_name}/pin": {
      "post": {
        "operationId": "toggle_favorite_tag_pin_api_favorite_tags__tag_name__pin_post",
        "parameters": [
          {
            "in": "path",
            "name": "tag_name",
            "required": true,
            "schema": {
              "title": "Tag Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Favorite Tag Pin"
      }
    },
    "/api/favorites/batch": {
      "put": {
        "operationId": "update_favorites_batch_api_favorites_batch_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteBatchUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Favorites Batch"
      }
    },
...
    "/api/favorites/{file_id}": {
      "post": {
        "operationId": "toggle_favorite_api_favorites__file_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "title": "File Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Favorite"
      }
    },
    "/api/favorites/{file_id}/pin": {
      "post": {
        "operationId": "toggle_favorite_pin_api_favorites__file_id__pin_post",
        "parameters": [
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "title": "File Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Favorite Pin"
      }
    },
...
  }
}